kolibri: Two key swaps
Swap ' and _ to avoid thumb jumping when typing "I'm" and to to avoid
triggering tap+hold of _ when typing UPPER_CASE_MACRO names.

Swap Ins and Enter to make accidental Enter even less likely. Hitting it
on the inner column requires a deliberate movement. It also puts Enter
on the stronger finger.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>