kolibri: more keyswaps
'!' and '~' are commonly used in bigrams '!=' and '~=' in C-like
programming languages. Therefore move them back to the Num+Sym layer.
Instead '#' and '@' take the places of '<' and '>', as they are less
commonly used in symbol bigrams.

Swap '|' and '\'. It turns out that '|' is used much more frequently, as
pipe in the shell or logical/binary OR in programming, so it should have
the less awkward key.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>