kolibri: Introduce numpad-style symbol layer
This is an alternative symbol layer that has a numpad on the right hand
and rearranges other symbols. It tries to maintain some commonality with
the original symbol layer where practical. The numpad is more
comfortable for numerical data entry and matches muscle memory of users
acustomed to using a numpad.

Swap @ # on the base layer to match the number keys on the numpad layer.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>