kolibri: Add 36-key version
Two extra thumb keys are used to eliminate two bottom-row mod-taps and
to duplicate Del on the base layer and '.' on the symbol layer. Space
becomes single-function key.

The left-Alt modifier is moved to the outer left thumb key as one-shot
modifier, which is easy to combine with Shift and with symbol layer
keys. This eliminates two bottom-row mod-taps.

AltGr moves to the outer right thumb key, making Space a single-function
key with normal repeat and no risk of accidentally triggering AltGr
while typing fast.

The outer right thumb key also duplicates Del on the base layer and '.'
on the symbol layer. AltGr+'.' is only possible on the base layer.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>