kolibri: Move / to Sym, " to Base layer
" makes more sense on the base layer as it avoids thumb-jumping for any
quotation starting with a capital letter. / makes more sense on the Sym
layer for math, C-comments, etc.

? is still on the base layer but on a thumb key, sharing Shift. Since it
is rarely used it should not interfere with normal typing and the
Hold-preference of the Shift key.

The Sym layer variants are rearranged a bit to put / in a sensible
place, either next to the numpad or in its normal bottom-right location.
On the num-row layer, + is now in a comfortable place in the top row.

Also moved ; and : to be in the same places in all Sym layer variants. ;
in the top right (normal Colemak place) and : on the inner thumb key.
This makes the := bigram more comfortable and avoids tap-hold ambibuity
when typing ; <Enter>.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>