kolibri: Fix Media/Macro layers for lefty layouts
Media/Macro layers are the same for right and left-handed layouts. But
the layer switching keys that activate them are on the wrong thumbs.
Remap the layer switching keys to make the layers usable. Also add RAW
layout macros that don't mirror the thumb clusters of these two layers
on left-handed layouts.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>