kolibri: Redesigned Mouse layer
Made the mouse layer a locked layer that is toggled via the Fn layer. It
is under the Fn layer so the Fn layer can be used on top of it, and the
layer can be disabled the same way it was enabled.

Tab and Del are available on the right thumb, like on the base layer,
for convenience. Space is replaced with a second acceleration button.

Movement and scrolling in the right hand, analogous to the inverted-T
arrow keys, page-up/down and home/end on the Fn layer. All buttons are
on the left hand. TODO: Make a left-handed version.

Shift is in the normal position on a thumb. All other modifiers are
available in the bottom row on both hands.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>