kolibri: Chord to clear all keyboard state
Add a chord using the thumb on the Config layer to clear all keyboard
state. It works in all locked layer states, including the mouse layer.
The chord disables all locked layers and attempts to disable caps lock
if it is enabled. Disabling caps lock may fail or produce unexpected
results if the host remaps the caps lock key.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>