kolibri: Simplify CapsLock and layer lock
Single tap of Fn+Shift enable CapsWord. Double tap enable CapsLock.
CapsWord and CapsLock can be combined.

Enable FnLock with a simple one-handed tap on the Fn layer. Tapping the
same key on the locked layer disables it. Could probably use that key
for something else on the locked layer, since disable can be done with
Fn+Sym.

Enable the Mouse layer with a simple one-handed tap in the Fn layer.

Moved KC_CALC and KC_PAUS to the MEDIA layer to make room.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>