Update Mantis v0.2
Mantis v0.2 changes some key positions and simplifies the matrix a bit.
Also revamped the default layout. Trying out a number row instead of a
numpad style Symbol layer. Changed the Function layer with F1-F10 in the
top row and adjusting the layout of the arrow key cluster accordingly.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>