mantis: Optimize symbol layer
Duplicating < and > on the base and symbol layers is not useful. They
are more useful on the symbol layer for bigrams like >= and ->. Use key
overrides to replace them with @ and # on the base layer. Instead
duplicate - and . on both layers to reduces the need of layer switching.

Reduce same-finger bigrams on the right pinky by moving : and = to other
fingers.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>