mantis: Rework indicator LEDs
Use HSV so that the same CIE1931_CURVE gets applied to the brightness
the same way as for RGB animations.

Add indicators for RGB keys in rainbow colors. Decrementing keys are
dimmer. Keys change color to reflect HSV settings. When Shift is held,
the dimming of keys is inverted to indicate the change in behaviour.

Signed-off-by: Felix Kuehling <felix.kuehling@gmail.com>